renaming utils package to give more context
